backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
psycopg2-binary
redis
scikit-learn
jinja2
requests
httpx
numpy
textblob
python-multipart
authlib
